@@ -1,15 +1,50 @@
+buildscript {
+    repositories {
+        google()
+        jcenter()
+        mavenCentral()
+    }
+
+//    dependencies {
+//        classpath 'me.tatarka:gradle-retrolambda:3.7.0'
+//    }
+}
+
+// Required because retrolambda is on maven central
+repositories {
+    google()
+    jcenter()
+    mavenCentral()
+//    maven {
+//        url 'https://maven.google.com'
+//    }
+//    maven {
+//        url 'https://oss.sonatype.org/content/repositories/snapshots/'
+//    }
+    maven { url 'https://jitpack.io' }
+}
+
 apply plugin: 'com.android.application'
+//apply plugin: 'me.tatarka.retrolambda'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 28
+    buildToolsVersion '29.0.0'
     defaultConfig {
-        applicationId "org.rares.miner49er"
-        minSdkVersion 14
-        targetSdkVersion 27
+        applicationId 'org.rares.miner49er'
+        minSdkVersion rootProject.minSdkVersion
+        targetSdkVersion rootProject.targetSdkVersion
         versionCode 1
-        versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        versionName '1.0'
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
+        multiDexEnabled true
+
+        javaCompileOptions {
+            annotationProcessorOptions {
+//                classNames ["butterknife.compiler.ButterKnifeProcessor"]
+                includeCompileClasspath false
+            }
+        }
     }
     buildTypes {
         release {
